+++ patches/src/main/kotlin/app/morphe/patches/protonvpn/delay/Fingerprints.kt
@@ -1,0 +1,11 @@
+package app.morphe.patches.protonvpn.delay
+
+import app.morphe.patcher.Fingerprint
+
+object GetLongDelayFingerprint : Fingerprint (
+    custom = { method, _ -> method.name == "getChangeServerLongDelayInSeconds" }
+)
+
+object GetShortDelayFingerprint : Fingerprint (
+    custom  = { method, _ -> method.name == "getChangeServerShortDelayInSeconds" }
+)